+++ src/lib/utils/aria.js
@@ -1,0 +1,14 @@
+export const keyCodes = {
+  backspace: 8,
+  tab: 9,
+  enter: 13,
+  esc: 27,
+  pageUp: 33,
+  pageDown: 34,
+  end: 35,
+  home: 36,
+  left: 37,
+  up: 38,
+  right: 39,
+  down: 40,
+};