+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # state-template - Change Log
 
+## v0.4.3 (2019-08-09)
+
+- (chore) update react-styleguidist v9.1.13
+- (chore) update dev dependencies
+- (feat) pass `cellEdit` in Table to react-bootstrap-table-next
+- (feat) allow ref to inner table and hide pagination in Table
+- (feat) add FieldComboBox component for accessible combo box field
+- (feat) add options for yes/no fields in `fieldOptions` util
+- (fix) Table header now pushes menu to right when hiding search
+- (fix) FieldSelect height consistent with other fields / use scss variables
+- (refactor) use `componentDidUpdate` in place of deprecated `componentWillReceiveProps`
+
 ## v0.4.2 (2019-08-02)
 
 - (chore) update dependency versions